transcode: add advanced options for deinterlacing
This patch exposes additional configuration options for the
deinterlace_vaapi (hardware) and yadif (software) deinterlace filters:
* Deinterlace rate type (rate): frame or field
* Deinterlace fields only (auto): only deinterlace interlaced fields
* VAAPI Deinterlace mode (mode): Bob, Weave, MADI, MCDI (for VAAPI only)

These options allow for the transcode deinterlace config options to be
fine-tuned. Most noteably the deinterlace filters can now be confugured
with field-rate deinterlacing, which causes (for example) 25 fps interlaced
input at a 90 kHz timebase to produce 50 fps output with a 180 kHz timebase.

To maintain MPEG-TS compliance, the output timebase is fixed at 90 kHz,
with packet timestamps rescaled accordingly before shipping to the muxer.
This is done dynamically using libav’s av_packet_rescale_ts() function,
with the timebase taken from the final filter in the chain (via
AVFilterContext) and rescaled to the timebase of the output
AVCodecContext. This approach is robust against future changes to the
filter configuration, including different combinations of deinterlace
options, and supports fractional frame rates.

When field-rate deinterlacing is selected, this produces frames with
correct timing for 50 fps playback in a 90 kHz container, ensuring that
the transcoded output stream preserves the intended cadence and temporal
fidelity of the original interlaced source.