transcode: adjust framerate duration and pts for VAAPI field-based deinterlacers
When using a field-based VAAPI de-interlacer, ffmpeg (libav) will double the
output frame rate and timebase. For the frames themselves avlib therefore also
doubles the pts and duration of the decoded output.

  Ref: ffmpeg-6.1.1/libavfilter/vf_deinterlace_vaapi.c#L169-L172
       ffmpeg-6.1.1/libavfilter/vf_deinterlace_vaapi.c#L296

The result of all this is that (for example) as 25 fps interlaced stream with a
90Hz timebase is decoded to a 50fps stream but with a 180Hz timebase.

When transcoding we want to stay with a 90Hz time base in order for the encoder
to receive an MPEG-TS compliant stream.

Therefore, we don't touch the timebase of the output stream nor do we fiddle with
the timebase of the packets or frames themselves. Instead, we halve the pts and
duration values on the frames themselves before they are sent along to the
decoder. As a result, the frames have correctly stamped PTS values and durations
required for 50fps playback within a 90Hz timebase container. The frame rate of
the output stream is then doubled to keep this inline with the reality of the new
frame durations and pts values within the stream.

This all results in far smoother playback (for example 50p instead of 25p) since
each field returned by the decoder becomes its own frame for encoding.